Allow providers to validate and refresh sessions
This change allows authentication providers to implement two new types
of behavior: expiring a session and refreshing a session. Both changes
were made with the intent of supporting upcoming changes in the OAuth2
providers, but might find use in other types of authentication providers
as well.

The use case for expiration is that OAuth2 authentication results in an
access token that often has an expiry date. Using the access token
to make requests beyond that date will result in an error response. The
auth middleware has the opportunity to detect an invalid token early and
see the user replaces it, pre-empting a failed request later and so
resulting in a better user experience.

The use case for refreshing a session is that OAuth2 also provides a
mechanism for obtaining a new access token without involving the user,
by using a refresh token. When we obbtain a new access token this way we
do need to update the session to store it, and this commit makes that
possible.

To maintain backwards compatibility we add a default behavior to all
existing providers that makes them behave like before this change. The
default behavior will never cause sessions to be expired or refreshed.